pyproject.toml
[tool.poetry]
name = "heatcontrol"
version = "0.6.0"
description = ""
authors = ["Rami Rahikkala <rami.rahikkala@gmail.fi>"]
readme = "README.md"
packages = [{include = "heatcontrol"}]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.2"
schedule = "^1.1.0"
rpi-gpio = "^0.7.1"
mock-gpio = "^0.1.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
spot_price_control = 'heatcontrol:main'
